conv: Move some generic helpers into headers.
